refactor: move tenor to provider
